Extract Dor::PublicDescMetadataService from Dor::Describable
