+++ lib/dor-services.rb
@@ -144,6 +144,7 @@ module Dor
   autoload :MergeService,              'dor/services/merge_service'
   autoload :ResetWorkspaceService,     'dor/services/reset_workspace_service'
   autoload :CleanupResetService,       'dor/services/cleanup_reset_service'
+  autoload :PublicDescMetadataService, 'dor/services/public_desc_metadata_service'
 
   # Workflow Classes
   module Workflow
