+++ lib/dor/models/concerns/describable.rb
@@ -46,8 +46,12 @@ module Dor
     # @raise [CrosswalkError] Raises an Exception if the generated DC is empty or has no children
     # @return [Nokogiri::Doc] the DublinCore XML document object
     def generate_dublin_core(include_collection_as_related_item: true)
-      desc_md = descMetadata.ng_xml.dup(1)
-      add_collection_reference(desc_md) if include_collection_as_related_item
+      desc_md = if include_collection_as_related_item
+                  Nokogiri::XML(generate_public_desc_md(include_access_conditions: false))
+                else
+                  descMetadata.ng_xml
+                end
+
       dc_doc = MODS_TO_DC_XSLT.transform(desc_md)
       dc_doc.xpath('/oai_dc:dc/*[count(text()) = 0]', oai_dc: XMLNS_OAI_DC).remove # Remove empty nodes
       raise CrosswalkError, "Dor::Item#generate_dublin_core produced incorrect xml (no root):\n#{dc_doc.to_xml}" if dc_doc.root.nil?
@@ -56,146 +60,8 @@ module Dor
     end
 
     # @return [String] Public descriptive medatada XML
-    def generate_public_desc_md
-      doc = descMetadata.ng_xml.dup(1)
-      add_collection_reference(doc)
-      add_access_conditions(doc)
-      add_constituent_relations(doc)
-      doc.xpath('//comment()').remove
-      new_doc = Nokogiri::XML(doc.to_xml) { |x| x.noblanks }
-      new_doc.encoding = 'UTF-8'
-      new_doc.to_xml
-    end
-
-    # Create MODS accessCondition statements from rightsMetadata
-    # @param [Nokogiri::XML::Document] doc Document representing the descriptiveMetadata of the object
-    # @note this method modifies the passed in doc
-    def add_access_conditions(doc)
-      # clear out any existing accessConditions
-      doc.xpath('//mods:accessCondition', 'mods' => 'http://www.loc.gov/mods/v3').each {|n| n.remove}
-      rights = datastreams['rightsMetadata'].ng_xml
-
-      rights.xpath('//use/human[@type="useAndReproduction"]').each do |use|
-        txt = use.text.strip
-        next if txt.empty?
-        doc.root.element_children.last.add_next_sibling doc.create_element('accessCondition', txt, :type => 'useAndReproduction')
-      end
-      rights.xpath('//copyright/human[@type="copyright"]').each do |cr|
-        txt = cr.text.strip
-        next if txt.empty?
-        doc.root.element_children.last.add_next_sibling doc.create_element('accessCondition', txt, :type => 'copyright')
-      end
-      rights.xpath("//use/machine[#{ci_compare('type', 'creativecommons')}]").each do |lic_type|
-        next if lic_type.text =~ /none/i
-        lic_text = rights.at_xpath("//use/human[#{ci_compare('type', 'creativecommons')}]").text.strip
-        next if lic_text.empty?
-        new_text = "CC #{lic_type.text}: #{lic_text}"
-        doc.root.element_children.last.add_next_sibling doc.create_element('accessCondition', new_text, :type => 'license')
-      end
-      rights.xpath("//use/machine[#{ci_compare('type', 'opendatacommons')}]").each do |lic_type|
-        next if lic_type.text =~ /none/i
-        lic_text = rights.at_xpath("//use/human[#{ci_compare('type', 'opendatacommons')}]").text.strip
-        next if lic_text.empty?
-        new_text = "ODC #{lic_type.text}: #{lic_text}"
-        doc.root.element_children.last.add_next_sibling doc.create_element('accessCondition', new_text, :type => 'license')
-      end
-    end
-
-    # Remove existing relatedItem entries for collections from descMetadata
-    def remove_related_item_nodes_for_collections(doc)
-      doc.search('/mods:mods/mods:relatedItem[@type="host"]/mods:typeOfResource[@collection=\'yes\']', 'mods' => 'http://www.loc.gov/mods/v3').each do |node|
-        node.parent.remove
-      end
-    end
-
-    def add_related_item_node_for_collection(doc, collection_druid)
-      begin
-        collection_obj = Dor.find(collection_druid)
-      rescue ActiveFedora::ObjectNotFoundError
-        return nil
-      end
-
-      title_node         = Nokogiri::XML::Node.new('title', doc)
-      title_node.content = Dor::Describable.get_collection_title(collection_obj)
-
-      title_info_node = Nokogiri::XML::Node.new('titleInfo', doc)
-      title_info_node.add_child(title_node)
-
-      # e.g.:
-      #   <location>
-      #     <url>http://purl.stanford.edu/rh056sr3313</url>
-      #   </location>
-      loc_node = doc.create_element('location')
-      url_node = doc.create_element('url')
-      url_node.content = "https://#{Dor::Config.stacks.document_cache_host}/#{collection_druid.split(':').last}"
-      loc_node << url_node
-
-      type_node = Nokogiri::XML::Node.new('typeOfResource', doc)
-      type_node['collection'] = 'yes'
-
-      related_item_node = Nokogiri::XML::Node.new('relatedItem', doc)
-      related_item_node['type'] = 'host'
-
-      related_item_node.add_child(title_info_node)
-      related_item_node.add_child(loc_node)
-      related_item_node.add_child(type_node)
-
-      doc.root.add_child(related_item_node)
-    end
-
-    # Adds to desc metadata a relatedItem with information about the collection this object belongs to.
-    # For use in published mods and mods-to-DC conversion.
-    # @param [Nokogiri::XML::Document] doc A copy of the descriptiveMetadata of the object, to be modified
-    # @return [Void]
-    # @note this method modifies the passed in doc
-    def add_collection_reference(doc)
-      return unless methods.include? :public_relationships
-      collections = public_relationships.search('//rdf:RDF/rdf:Description/fedora:isMemberOfCollection',
-                                       'fedora' => 'info:fedora/fedora-system:def/relations-external#',
-                                       'rdf' => 'http://www.w3.org/1999/02/22-rdf-syntax-ns#')
-      return if collections.empty?
-
-      remove_related_item_nodes_for_collections(doc)
-
-      collections.each do |collection_node|
-        collection_druid = collection_node['rdf:resource'].gsub('info:fedora/', '')
-        add_related_item_node_for_collection(doc, collection_druid)
-      end
-    end
-
-    # expand constituent relations into relatedItem references -- see JUMBO-18
-    # @param [Nokogiri::XML] doc public MODS XML being built
-    # @return [Void]
-    def add_constituent_relations(doc)
-      public_relationships.search('//rdf:RDF/rdf:Description/fedora:isConstituentOf',
-                                       'fedora' => 'info:fedora/fedora-system:def/relations-external#',
-                                       'rdf' => 'http://www.w3.org/1999/02/22-rdf-syntax-ns#' ).each do |parent|
-        # fetch the parent object to get title
-        druid = parent['rdf:resource'].gsub(/^info:fedora\//, '')
-        parent_item = Dor.find(druid)
-
-        # create the MODS relation
-        relatedItem = doc.create_element 'relatedItem'
-        relatedItem['type'] = 'host'
-        relatedItem['displayLabel'] = 'Appears in'
-
-        # load the title from the parent's DC.title
-        titleInfo = doc.create_element 'titleInfo'
-        title = doc.create_element 'title'
-        title.content = Dor::Describable.get_collection_title(parent_item)
-        titleInfo << title
-        relatedItem << titleInfo
-
-        # point to the PURL for the parent
-        location = doc.create_element 'location'
-        url = doc.create_element 'url'
-        url.content = "http://#{Dor::Config.stacks.document_cache_host}/#{druid.split(':').last}"
-        location << url
-        relatedItem << location
-
-        # finish up by adding relation to public MODS
-        doc.root << relatedItem
-      end
+    def generate_public_desc_md(**options)
+      PublicDescMetadataService.new(self).to_xml(**options)
     end
 
     def to_solr(solr_doc = {}, *args)
@@ -296,17 +162,6 @@ module Dor
 
     def full_title
       stanford_mods.sw_title_display
-    end
-
-    private
-
-    # Builds case-insensitive xpath translate function call that will match the attribute to a value
-    def ci_compare(attribute, value)
-      "translate(
-        @#{attribute},
-        'ABCDEFGHIJKLMNOPQRSTUVWXYZ',
-        'abcdefghijklmnopqrstuvwxyz'
-       ) = '#{value}' "
     end
   end
 end
